реализованы вывод времени в привычном формате и обработка исключений
